Edited Modal to ensure modularity of code
