+++ src/components/styles/ItineraryPage.css
@@ -1,53 +1,67 @@
-.itinerary-page {
-    padding: 20px;
-    padding-top: 100px; 
-}
-  
-.itinerary-page h1 {
-    font-size: 2em;
-    margin-bottom: 10px;
-}
-  
-.itinerary-page p {
-    font-size: 1.2em;
-    margin-bottom: 20px;
-}
-  
-.meetings-list {
-    display: flex;
-    flex-direction: column;
-}
-  
-.day-meetings {
-    margin-bottom: 20px;
-}
-  
-.day-meetings h2 {
-    font-size: 1.5em;
-    margin-bottom: 10px;
-}
-  
-.day-meetings ul {
-    list-style: none;
-    padding: 0;
-}
-  
-.day-meetings li {
-    padding: 5px 0;
-}
-  
-.edit-itinerary-button {
-    background-color: #007bff;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    margin-top: 20px;
-    font-size: 1em;
-    cursor: pointer;
-    border-radius: 5px;
-    transition: background-color 0.3s ease;
-}
-  
-.edit-itinerary-button:hover {
-    background-color: #0056b3;
-}
+    
+    
+    .itinerary-page {
+        padding: 20px;
+        padding-top: 100px; 
+    }
+    
+    .itinerary-page h1 {
+        font-size: 2em;
+        margin-bottom: 10px;
+    }
+    
+    .itinerary-page p {
+        font-size: 1.2em;
+        margin-bottom: 20px;
+    }
+    
+    .meetings-list {
+        display: flex;
+        flex-direction: column;
+    }
+    
+    .day-meetings {
+        margin-bottom: 20px;
+    }
+    
+    .day-meetings h2 {
+        font-size: 1.5em;
+        margin-bottom: 10px;
+    }
+    
+    .day-meetings ul {
+        list-style: none;
+        padding: 0;
+    }
+    
+    .day-meetings li {
+        padding: 5px 0;
+    }
+    
+    .edit-itinerary-button {
+        background-color: #007bff;
+        color: white;
+        border: none;
+        padding: 10px 20px;
+        margin-top: 20px;
+        font-size: 1em;
+        cursor: pointer;
+        border-radius: 5px;
+        transition: background-color 0.3s ease;
+    }
+    
+    .edit-itinerary-button:hover {
+        background-color: #0056b3;
+    }
+
+    .edit-expenses-button {
+        background-color: purple;
+        color: white;
+        border: none;
+        padding: 10px 20px;
+        margin-top: 20px;
+        font-size: 1em;
+        cursor: pointer;
+        border-radius: 5px;
+        transition: background-color 0.3s ease;
+    }
